+++ config/chains/mainnet_v1.exs
@@ -30,5 +30,3 @@ config :godwoken_explorer,
   ckb_token_script_hash: "0x0000000000000000000000000000000000000000000000000000000000000000",
   eth_addr_reg_id: "0x2",
   polyjuice_creator_id: 4
-
-config :godwoken_explorer, :sourcify, chain_id: "71402"
